move MediaTypeWriter to streamformatter package
